+++ MestreDosCodigosDotNet/ExercicioPOO_3_XUnitTests/ContaEspecialTests.cs
@@ -6,15 +6,15 @@ namespace ExercicioPOO_3_XUnitTests
 {
     public class ContaEspecialTests
     {
-        private const Double SALDO_INICIAL = 100.0;
-        private const Double LIMITE = 200;
+        private const decimal SALDO_INICIAL = 100m;
+        private const decimal LIMITE = 200m;
         private const int NUMERO_CONTA = 1;
 
         [Fact]
         public void ContaEspecial_Sacar_DeveAtualizarSaldo()
         {
             // Arrange
-            var valorsaque = 90.0;
+            var valorsaque = 90m;
             var saldoAtualizado = Math.Round(SALDO_INICIAL - valorsaque, 2);
             var contaEspecial = PegarContaEspecial();
 
@@ -29,7 +29,7 @@ namespace ExercicioPOO_3_XUnitTests
         public void ContaEspecial_Sacar_NaoDeveAtualizarSaldo()
         {
             // Arrange
-            var valorsaque = 350;
+            var valorsaque = 350m;
             var contaEspecial = PegarContaEspecial();
 
             // Act
@@ -46,7 +46,7 @@ namespace ExercicioPOO_3_XUnitTests
             var contaEspecial = PegarContaEspecial();
 
             // Act
-            contaEspecial.Depositar(-1);
+            contaEspecial.Depositar(-1m);
 
             // Assert
             Assert.Equal(SALDO_INICIAL, contaEspecial.Saldo);
@@ -56,7 +56,7 @@ namespace ExercicioPOO_3_XUnitTests
         public void ContaEspecial_Depositar_DeveAtualizarSaldo()
         {
             // Arrange
-            var valorDeposito = 100;
+            var valorDeposito = 100m;
             var saldoAtualizado = Math.Round(SALDO_INICIAL + valorDeposito, 2);
             var contaEspecial = PegarContaEspecial();
